+++ packages/tezos-ts/src/contract/interface.ts
@@ -1,118 +1,7 @@
 import { Schema } from '@tezos-ts/michelson-encoder';
-import { Operation } from './operations';
-import { OperationObject } from '@tezos-ts/rpc';
+import { Operation } from '../operations';
 import { Contract } from './contract';
-
-/**
- * @description Parameters for originate method
- */
-export interface OriginateParams {
-  balance?: string;
-  code: string | object[];
-  init: string | object;
-  spendable?: boolean;
-  delegatable?: boolean;
-  delegate?: string;
-  fee?: number;
-  gasLimit?: number;
-  storageLimit?: number;
-}
-
-/**
- * @description RPC origination operation
- */
-export interface OriginationOperation {
-  kind: 'origination';
-  fee: number;
-  gas_limit: number;
-  storage_limit: number;
-  balance: string;
-  manager_pubkey: string;
-  spendable: boolean;
-  delegatable: boolean;
-  delegate?: string;
-  script: {
-    code: any;
-    storage: any;
-  };
-}
-
-/**
- * @description RPC reveal operation
- */
-export interface RevealOperation {
-  kind: 'reveal';
-  fee: number;
-  public_key: string;
-  source: string;
-  gas_limit: number;
-  storage_limit: number;
-}
-
-/**
- * @description Result of a forge operation contains the operation plus its encoded version
- */
-export interface ForgedBytes {
-  opbytes: string;
-  opOb: OperationObject;
-  counter: number;
-}
-
-/**
- * @description Parameters for setDelegate and registerDelegate method
- */
-export interface DelegateParams {
-  source: string;
-  delegate: string;
-  fee?: number;
-  gasLimit?: number;
-  storageLimit?: number;
-}
-
-/**
- * @description RPC delegation operation
- */
-export interface DelegateOperation {
-  kind: 'delegation';
-  source?: string;
-  fee: number;
-  gas_limit: number;
-  storage_limit: number;
-  delegate: string;
-}
-
-/**
- * @description Parameters for transfer method
- */
-export interface TransferParams {
-  to: string;
-  source?: string;
-  amount: number;
-  fee?: number;
-  parameter?: string;
-  gasLimit?: number;
-  storageLimit?: number;
-  mutez?: boolean;
-  rawParam?: boolean;
-}
-
-/**
- * @description RPC transfer operation
- */
-export interface TransferOperation {
-  kind: 'transaction';
-  fee: number;
-  gas_limit: number;
-  storage_limit: number;
-  amount: string;
-  destination: string;
-  parameters?: any;
-}
-
-export type PrepareOperationParams = {
-  operation: OriginationOperation | TransferOperation | DelegateOperation | RevealOperation;
-  source?: string;
-};
+import { DelegateParams, OriginateParams, TransferParams } from '../types';
 
 export type ContractSchema = Schema | unknown;
