+++ packages/tezos-ts/src/contract/rpc-contract-provider.ts
@@ -1,34 +1,25 @@
+import { ParameterSchema, Schema } from '@tezos-ts/michelson-encoder';
+import { ml2mic, sexp2mic } from '@tezos-ts/utils';
+import { DEFAULT_FEE, DEFAULT_GAS_LIMIT, DEFAULT_STORAGE_LIMIT } from '../constants';
+import { Context } from '../context';
+import { format } from '../format';
+import { OperationEmitter } from '../operations/operation-emitter';
+import { Operation } from '../operations/operations';
+import { Contract } from './contract';
+import { ContractProvider, ContractSchema } from './interface';
 import {
-  ContractProvider,
-  ContractSchema,
   OriginateParams,
   OriginationOperation,
-  PrepareOperationParams,
+  DelegateParams,
+  DelegateOperation,
   TransferParams,
   TransferOperation,
-  DelegateOperation,
-  DelegateParams,
-  ForgedBytes,
-  RevealOperation,
-} from './interface';
-import { Schema, ParameterSchema } from '@tezos-ts/michelson-encoder';
-import { ml2mic, sexp2mic } from '@tezos-ts/utils';
-import { format } from '../format';
-import { Context } from '../context';
-import { Operation } from './operations';
-import { BlockResponse } from '@tezos-ts/rpc';
-import { Contract } from './contract';
-import { DEFAULT_FEE, DEFAULT_GAS_LIMIT, DEFAULT_STORAGE_LIMIT } from './constants';
+  ActivateOperation,
+} from '../operations/types';
 
-export class RpcContractProvider implements ContractProvider {
-  constructor(private context: Context) {}
-
-  get rpc() {
-    return this.context.rpc;
-  }
-
-  get signer() {
-    return this.context.signer;
+export class RpcContractProvider extends OperationEmitter implements ContractProvider {
+  constructor(context: Context) {
+    super(context);
   }
 
   /**
@@ -84,167 +75,6 @@ export class RpcContractProvider implements ContractProvider {
     const val = await this.rpc.getBigMapKey(contract, encodedKey);
 
     return contractSchema.ExecuteOnBigMapValue(val) as T; // Cast into T because only the caller can know the true type of the storage
-  }
-
-  private async prepareOperation({
-    operation,
-    source,
-  }: PrepareOperationParams): Promise<ForgedBytes> {
-    let counter;
-    const counters: { [key: string]: number } = {};
-    const promises: any[] = [];
-    let requiresReveal = false;
-    let ops: PrepareOperationParams['operation'][] = [];
-    let head: BlockResponse;
-
-    promises.push(this.rpc.getBlockHeader());
-
-    if (Array.isArray(operation)) {
-      ops = [...operation];
-    } else {
-      ops = [operation];
-    }
-
-    const publicKeyHash = source || (await this.signer.publicKeyHash());
-
-    for (let i = 0; i < ops.length; i++) {
-      if (['transaction', 'origination', 'delegation'].includes(ops[i].kind)) {
-        requiresReveal = true;
-        const { counter } = await this.rpc.getContract(publicKeyHash);
-        promises.push(Promise.resolve(counter));
-        promises.push(this.rpc.getManagerKey(publicKeyHash));
-        break;
-      }
-    }
-
-    promises.push(this.rpc.getBlockMetadata());
-
-    return Promise.all(promises).then(
-      async ([header, headCounter, manager, metadata]: any[]): Promise<any> => {
-        head = header;
-
-        const managerKey = manager.key;
-        if (requiresReveal && !managerKey) {
-          const reveal: RevealOperation = {
-            kind: 'reveal',
-            fee: DEFAULT_FEE.REVEAL,
-            public_key: await this.signer.publicKey(),
-            source: publicKeyHash,
-            gas_limit: DEFAULT_GAS_LIMIT.REVEAL,
-            storage_limit: DEFAULT_STORAGE_LIMIT.REVEAL,
-          };
-
-          ops.unshift(reveal);
-        }
-
-        counter = parseInt(headCounter, 10);
-        if (!counters[publicKeyHash] || counters[publicKeyHash] < counter) {
-          counters[publicKeyHash] = counter;
-        }
-
-        const constructOps = (cOps: PrepareOperationParams['operation'][]) =>
-          cOps.map((op: any) => {
-            // @ts-ignore
-            const constructedOp = { ...op };
-            if (['transaction', 'origination', 'delegation'].includes(op.kind)) {
-              if (typeof op.source === 'undefined') {
-                constructedOp.source = publicKeyHash;
-              }
-            }
-            if (['reveal', 'transaction', 'origination', 'delegation'].includes(op.kind)) {
-              if (typeof op.fee === 'undefined') {
-                constructedOp.fee = '0';
-              } else {
-                constructedOp.fee = `${op.fee}`;
-              }
-              if (typeof op.gas_limit === 'undefined') {
-                constructedOp.gas_limit = '0';
-              } else {
-                constructedOp.gas_limit = `${op.gas_limit}`;
-              }
-              if (typeof op.storage_limit === 'undefined') {
-                constructedOp.storage_limit = '0';
-              } else {
-                constructedOp.storage_limit = `${op.storage_limit}`;
-              }
-              if (typeof op.balance !== 'undefined') {
-                constructedOp.balance = `${constructedOp.balance}`;
-              }
-              if (typeof op.amount !== 'undefined') {
-                constructedOp.amount = `${constructedOp.amount}`;
-              }
-              const opCounter = ++counters[publicKeyHash];
-              constructedOp.counter = `${opCounter}`;
-            }
-            return constructedOp;
-          });
-
-        const branch = head.hash;
-        const contents = constructOps(ops);
-        const protocol = metadata.nextProtocol;
-
-        let remoteForgedBytes = await this.rpc.forgeOperations({ branch, contents });
-
-        return {
-          opbytes: remoteForgedBytes,
-          opOb: {
-            branch,
-            contents,
-            protocol,
-          },
-          counter,
-        };
-      }
-    );
-  }
-
-  private async signAndInject(forgedBytes: ForgedBytes) {
-    const signed = await this.signer.sign(forgedBytes.opbytes, new Uint8Array([3]));
-    forgedBytes.opbytes = signed.sbytes;
-    forgedBytes.opOb.signature = signed.prefixSig;
-
-    const opResponse: any[] = [];
-    let errors: any[] = [];
-
-    const results = await this.rpc.preapplyOperations([forgedBytes.opOb]);
-
-    if (!Array.isArray(results)) {
-      throw new Error(`RPC Fail: ${JSON.stringify(results)}`);
-    }
-    for (let i = 0; i < results.length; i++) {
-      for (let j = 0; j < results[i].contents.length; j++) {
-        opResponse.push(results[i].contents[j]);
-        const content = results[i].contents[j];
-        if (
-          'metadata' in content &&
-          typeof content.metadata.operation_result !== 'undefined' &&
-          content.metadata.operation_result.status === 'failed'
-        ) {
-          errors = errors.concat(content.metadata.operation_result.errors);
-        }
-      }
-    }
-
-    if (errors.length) {
-      // @ts-ignore
-      throw new Error(JSON.stringify({ error: 'Operation Failed', errors }));
-    }
-
-    return new Operation(
-      await this.rpc.injectOperation(forgedBytes.opbytes),
-      forgedBytes,
-      opResponse,
-      this.context.clone()
-    );
-  }
-
-  async at(address: string): Promise<any> {
-    let script = await this.rpc.getScript(address);
-
-    let contractSchema: Schema = Schema.fromRPCResponse({ script });
-    let parameterSchema: ParameterSchema = ParameterSchema.fromRPCResponse({ script });
-
-    return new Contract(address, contractSchema, parameterSchema, this);
   }
 
   /**
@@ -379,5 +209,14 @@ export class RpcContractProvider implements ContractProvider {
 
     const opBytes = await this.prepareOperation({ operation, source });
     return this.signAndInject(opBytes);
+  }
+
+  async at(address: string): Promise<any> {
+    let script = await this.rpc.getScript(address);
+
+    let contractSchema: Schema = Schema.fromRPCResponse({ script });
+    let parameterSchema: ParameterSchema = ParameterSchema.fromRPCResponse({ script });
+
+    return new Contract(address, contractSchema, parameterSchema, this);
   }
 }
