+++ packages/tezos-ts/src/tz/interface.ts
@@ -1,4 +1,5 @@
-import BigNumber from "bignumber.js";
+import BigNumber from 'bignumber.js';
+import { Operation } from '../operations';
 
 export interface TzProvider {
   /**
@@ -12,4 +13,6 @@ export interface TzProvider {
    * @param address Tezos address you want to get the delegate for (eg tz1...)
    */
   getDelegate(address: string): Promise<string | null>;
+
+  activate(pkh: string, secret: string): Promise<Operation>;
 }
