+++ packages/tezos-ts/src/tz/rpc-tz-provider.ts
@@ -2,12 +2,13 @@ import { TzProvider } from './interface';
 import { RpcClient } from '@tezos-ts/rpc';
 import BigNumber from 'bignumber.js';
 import { Context } from '../context';
+import { ActivateOperation } from '../operations/types';
+import { Operation } from '../operations/operations';
+import { OperationEmitter } from '../operations/operation-emitter';
 
-export class RpcTzProvider implements TzProvider {
-  constructor(private context: Context) {}
-
-  get rpc() {
-    return this.context.rpc;
+export class RpcTzProvider extends OperationEmitter implements TzProvider {
+  constructor(context: Context) {
+    super(context);
   }
 
   async getBalance(address: string): Promise<BigNumber> {
@@ -16,5 +17,22 @@ export class RpcTzProvider implements TzProvider {
 
   async getDelegate(address: string): Promise<string | null> {
     return this.rpc.getDelegate(address);
+  }
+
+  async activate(pkh: string, secret: string) {
+    const operation: ActivateOperation = {
+      kind: 'activate_account',
+      pkh,
+      secret,
+    };
+
+    const forgedBytes = await this.prepareOperation({ operation: [operation], source: pkh });
+    const bytes = `${forgedBytes.opbytes}00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000`;
+    return new Operation(
+      await this.rpc.injectOperation(bytes),
+      { ...forgedBytes, opbytes: bytes },
+      [],
+      this.context.clone()
+    );
   }
 }
